Modularize L10N JSON File - Security
